requirements.txt
mysql-connector-python
numpy
python-dotenv
sqlalchemy
